Vino-Model-Streamlit/requirements.txt
pandas
streamlit
torch
torchvision
tensorflow
ultralytics
opencv-python
numpy
Pillow
opencv-python-headless

